Add modelling prototype for reading from hord
This lets me define a model for a hord, including casting to/from the file. Means that the casting logic can be removed from the application side and delegated to wisdomhord. Next is to define a bunch of standard data types I want, rather than writing the lambdas myself for each model